refactor GoInspectionTestCase, make doTestWithDirectory() private and change the base path of directory test cases to the directory
