+++ test/ro/redeul/google/go/inspection/ImportUnusedInspectionTest.java
@@ -2,8 +2,8 @@ package ro.redeul.google.go.inspection;
 
 public class ImportUnusedInspectionTest extends GoInspectionTestCase {
 
-    public void testSimple() throws Exception{ doTestWithDirectory(); }
-    public void testOnlyOneImport() throws Exception{ doTestWithDirectory(); }
+    public void testSimple() throws Exception{ doTest(); }
+    public void testOnlyOneImport() throws Exception{ doTest(); }
 
     public void testBlankImport() throws Exception{ doTest(); }
     public void testMixedCaseImport() throws Exception{ doTest(); }
