+++ test/ro/redeul/google/go/inspection/UnresolvedSymbolsTest.java
@@ -65,7 +65,7 @@ public class UnresolvedSymbolsTest extends GoInspectionTestCase {
     }
 
     public void testIssue858() throws Exception {
-        doTestWithDirectory();
+        doTest();
     }
     
     public void testIssue979() throws Exception {
@@ -83,11 +83,11 @@ public class UnresolvedSymbolsTest extends GoInspectionTestCase {
 
     @Ignore("failing test")
     public void testGoTestFiles() throws Exception{
-        doTestWithDirectory();
+        doTest();
     }
 
     @Ignore("failing")
     public void testMethodFromAnotherPackage() throws Exception{
-        doTestWithDirectory();
+        doTest();
     }
 }
