+++ testdata/inspection/functionCall/issue875/p2/p2.go
@@ -1,7 +1,7 @@
 package p2
 
 import (
-	"issue875/p1"
+	"p1"
 )
 
 func f1() {
