+++ testdata/inspection/functionReturnParameter/io/Writer.go
@@ -1,4 +1,4 @@
-package "io"
+package io
 
 type Writer interface {
 	Write(p []byte) (n int, err error)
