+++ testdata/inspection/importPath/notFound/notFound.go
@@ -1,6 +1,6 @@
 package notFound
 
 import (
-	"notFound/package1"
+	"package1"
 	/*begin*/"package2"/*end.Package "package2" not found in project*/
 )
