--- testdata/inspection/importPath/self/p1.go
@@ -1,5 +1,0 @@
-package self
-
-import (
-	/*begin*/"self"/*end.You can not import current package "self"*/
-)
