+++ testdata/inspection/importPath/self/p1/p1.go
@@ -1,0 +1,5 @@
+package p1
+
+import (
+	/*begin*/"p1"/*end.You can not import current package "p1"*/
+)
