+++ testdata/inspection/importUnused/simple/simple.go
@@ -1,8 +1,8 @@
 package simple
 
 import (
-	"simple/p1"
-	/*begin*/"simple/p2"/*end.Unused import "simple/p2"|RemoveImportFix*/
+	"p1"
+	/*begin*/"p2"/*end.Unused import "p2"|RemoveImportFix*/
 )
 
 func main() {
