+++ testdata/inspection/unresolvedSymbols/goTestFiles/goTestFiles/goTestFiles_test.go
