+++ testdata/inspection/unresolvedSymbols/goTestFiles/goTestFiles/notTest.go
@@ -1,0 +1,5 @@
+package goTestFiles
+
+func bad(){
+	_=/*begin*/T1/*end.Unresolved symbol: 'T1'|CreateFunctionFix|CreateClosureFunctionFix*/(1)
+}
