+++ testdata/inspection/unresolvedSymbols/goTestFiles/goTestFiles/tt_test.go
