--- testdata/inspection/unresolvedSymbols/goTestFiles/notTest.go
@@ -1,5 +1,0 @@
-package goTestFiles
-
-func bad(){
-	_=/*begin*/T1/*end.Unresolved symbol: 'notExist1'*/(1)
-}
