+++ testdata/inspection/unresolvedSymbols/issue858/p2/p2.go
@@ -1,7 +1,7 @@
 package p2
 
 import (
-	"issue858/p1"
+	"p1"
 )
 
 func good() {
