+++ testdata/inspection/unusedVariable/issue438/issue438/demopack1.go
