+++ testdata/inspection/unusedVariable/issue438/issue438/demopack2.go
